backend/app/schemas/user.py
from pydantic import BaseModel, ConfigDict, EmailStr, Field


class userCreate(BaseModel):
    name: str = Field(..., min_length=1, max_length=120)
    email: EmailStr
    password: str = Field(..., min_length=8, max_length=128)
    company_code: str = Field(..., min_length=1, max_length=200)


class userLogin(BaseModel):
    email: EmailStr
    password: str = Field(..., min_length=1, max_length=128)


class userSimple(BaseModel):
    name: str = Field(..., min_length=1, max_length=120)
    email: EmailStr


class userResponse(userSimple):
    id: int

    model_config = ConfigDict(from_attributes=True)
