backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
passlib[argon2]>=1.7.4
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
psycopg2-binary>=2.9.9
slowapi>=0.1.9
email-validator>=2.0.0
